+++ R/global_vars.R
@@ -1,1 +1,1 @@
-globalVariables(c("index", "index.num", "y", "yhat", "week", "mday", "diff.median"))
+globalVariables(c("index", "index.num", "y", "yhat", "week", "mday", "diff.median", "."))